Database retrieval and updating, posting via http upon completing questionnaire
