+++ .idea/modules.xml
@@ -2,8 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/InspiringFutures.iml" filepath="$PROJECT_DIR$/InspiringFutures.iml" />
+      <module fileurl="file://$PROJECT_DIR$/InspiringFutures-client.iml" filepath="$PROJECT_DIR$/InspiringFutures-client.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/server/server.iml" filepath="$PROJECT_DIR$/server/server.iml" />
     </modules>
   </component>
 </project>
