+++ app/src/main/java/uk/ac/cam/crim/inspiringfutures/ESM/UncheckableRadioButton.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2017 Gideon Mills
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.ac.cam.crim.inspiringfutures.ESM;
+
+import android.content.Context;
+import android.support.v7.widget.AppCompatRadioButton;
+import android.util.AttributeSet;
+import android.widget.RadioGroup;
+
+/**
+ * <p> Created by  Gideon Mills on 25/07/2017 for InspiringFutures. </p>
+ */
+
+public class UncheckableRadioButton extends AppCompatRadioButton {
+
+    public UncheckableRadioButton(Context context) {
+        super(context);
+    }
+
+    public UncheckableRadioButton(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public UncheckableRadioButton(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+    }
+
+    @Override
+    public void toggle() {
+        if(isChecked()) {
+            if(getParent() instanceof RadioGroup) {
+                ((RadioGroup)getParent()).clearCheck();
+            }
+        } else {
+            setChecked(true);
+        }
+    }
+
+}
