--- app/src/main/java/uk/ac/cam/gsm31/inspiringfutures/util/InvalidTypeException.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2017 Gideon Mills
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.ac.cam.gsm31.inspiringfutures.util;
-
-/**
- * Indicates an invaid type was passed to putContentValues to be inserted into a ContentValues object
- */
-public class InvalidTypeException extends RuntimeException {
-    public InvalidTypeException(String s) {
-        super(s);
-    }
-}
