+++ app/src/main/res/drawable/ic_inspiringfutures_launcher.xml
@@ -1,0 +1,39 @@
+<!--
+  ~ Copyright 2017 Gideon Mills
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<vector android:height="24dp" android:viewportHeight="450.99994"
+    android:viewportWidth="450.99997" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillAlpha="1" android:fillColor="#bebebe"
+        android:pathData="M225.5,235.5m-200,0a200,200 0,1 1,400 0a200,200 0,1 1,-400 0"
+        android:strokeAlpha="1" android:strokeColor="#00000000"
+        android:strokeLineCap="butt" android:strokeLineJoin="round" android:strokeWidth="1"/>
+    <path android:fillAlpha="1" android:fillColor="#ffffff"
+        android:pathData="M225.5,225.5m-200,0a200,200 0,1 1,400 0a200,200 0,1 1,-400 0"
+        android:strokeAlpha="1" android:strokeColor="#00000000"
+        android:strokeLineCap="butt" android:strokeLineJoin="round" android:strokeWidth="1"/>
+    <path android:fillAlpha="1" android:fillColor="#f5f5f5"
+        android:pathData="m200.5,375.5 l0,-200 50,0 -50,-50 0,-50 0,0 50,0 173.6,173.5C413.2,340.9 340.8,413.3 249,424.1Z"
+        android:strokeAlpha="1" android:strokeColor="#00000000"
+        android:strokeLineCap="butt" android:strokeLineJoin="round" android:strokeWidth="1"/>
+    <path android:fillAlpha="1" android:fillColor="#2da8ff"
+        android:pathData="m200.5,375.5 l0,-200 100,0 0,50 -50,0 0,50 50,0 0,50 -50,0 0,50z"
+        android:strokeAlpha="1" android:strokeColor="#00000000"
+        android:strokeLineCap="butt" android:strokeLineJoin="miter" android:strokeWidth="1"/>
+    <path android:fillAlpha="1" android:fillColor="#2da8ff"
+        android:pathData="m200.5,125.5 l0,-50 50,0 0,50z"
+        android:strokeAlpha="1" android:strokeColor="#00000000"
+        android:strokeLineCap="butt" android:strokeLineJoin="miter" android:strokeWidth="1"/>
+</vector>
